Actions for F5,F7,F9, and mouseclick are now customizable
It is now possible to customize the functions of F5,F7,F9, and mouseclick in the config file as well as disable command execution and browser execution.  The actions can be text output or command execution.  For users concerned about the safety of the command execution features, disable-exec and disable-browser are available from the command line or from the [gource] section of the config file.